Added Jacobian of things that could be used in iteration
Very fast in C++ (less than a microsecond for 4 outputs in the Jacobian), but 10 microseconds in Python for 2. Need to find a way to make it closer to original speed. Perhaps pass in a buffer to be populated.  That should allow the copy to be avoided. It must be the Eigen arrays that are so slow.